fix(header): refactored header stories to include the js in the code example
